+++ circuit-foundation/src/commonMain/kotlin/com/slack/circuit/foundation/NavigableCircuitContent.kt
@@ -4,6 +4,8 @@ package com.slack.circuit.foundation
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.ContentTransform
+import androidx.compose.animation.EnterTransition
+import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.SizeTransform
 import androidx.compose.animation.core.CubicBezierEasing
 import androidx.compose.animation.core.LinearEasing
@@ -32,6 +34,7 @@ import com.slack.circuit.backstack.NavDecoration
 import com.slack.circuit.backstack.ProvidedValues
 import com.slack.circuit.backstack.isEmpty
 import com.slack.circuit.backstack.providedValuesForBackStack
+import com.slack.circuit.foundation.NavigatorDefaults.DefaultDecoration.backward
 import com.slack.circuit.retained.CanRetainChecker
 import com.slack.circuit.retained.LocalCanRetainChecker
 import com.slack.circuit.retained.LocalRetainedStateRegistry
@@ -238,35 +241,52 @@ public object NavigatorDefaults {
       val enterTransition =
         fadeIn(
           animationSpec =
-            tween(durationMillis = SHORT_DURATION, delayMillis = 50, easing = LinearEasing)
+            tween(
+              durationMillis = SHORT_DURATION,
+              delayMillis = if (sign > 0) 50 else 0,
+              easing = LinearEasing,
+            )
         ) +
           slideInHorizontally(
             initialOffsetX = { fullWidth -> (fullWidth / 10) * sign },
             animationSpec =
               tween(durationMillis = NORMAL_DURATION, easing = FastOutExtraSlowInEasing),
           ) +
-          expandHorizontally(
-            animationSpec =
-              tween(durationMillis = NORMAL_DURATION, easing = FastOutExtraSlowInEasing),
-            initialWidth = { (it * .9f).toInt() },
-            expandFrom = Alignment.Start,
-          )
+          if (sign > 0) {
+            expandHorizontally(
+              animationSpec =
+                tween(durationMillis = NORMAL_DURATION, easing = FastOutExtraSlowInEasing),
+              initialWidth = { (it * .9f).toInt() },
+              expandFrom = if (sign > 0) Alignment.Start else Alignment.End,
+            )
+          } else {
+            EnterTransition.None
+          }
 
       val exitTransition =
         fadeOut(
-          animationSpec = tween(durationMillis = NORMAL_DURATION, easing = AccelerateEasing)
+          animationSpec =
+            tween(
+              durationMillis = if (sign > 0) NORMAL_DURATION else SHORT_DURATION,
+              delayMillis = if (sign > 0) 0 else 50,
+              easing = AccelerateEasing,
+            )
         ) +
           slideOutHorizontally(
             targetOffsetX = { fullWidth -> (fullWidth / 10) * -sign },
             animationSpec =
               tween(durationMillis = NORMAL_DURATION, easing = FastOutExtraSlowInEasing),
           ) +
-          shrinkHorizontally(
-            animationSpec =
-              tween(durationMillis = NORMAL_DURATION, easing = FastOutExtraSlowInEasing),
-            targetWidth = { (it * .9f).toInt() },
-            shrinkTowards = Alignment.End,
-          )
+          if (sign > 0) {
+            shrinkHorizontally(
+              animationSpec =
+                tween(durationMillis = NORMAL_DURATION, easing = FastOutExtraSlowInEasing),
+              targetWidth = { (it * .9f).toInt() },
+              shrinkTowards = Alignment.End,
+            )
+          } else {
+            ExitTransition.None
+          }
 
       return enterTransition togetherWith exitTransition
     }
@@ -283,7 +303,8 @@ public object NavigatorDefaults {
         targetState = args,
         modifier = modifier,
         transitionSpec = {
-          // A transitionSpec should only use values passed into the `AnimatedContent`, to minimize
+          // A transitionSpec should only use values passed into the `AnimatedContent`, to
+          // minimize
           // the transitionSpec recomposing. The states are available as `targetState` and
           // `initialState`
           val diff = targetState.size - initialState.size