+++ circuitx/gesture-navigation/src/androidUnitTest/kotlin/com/slack/circuitx/gesturenavigation/GestureNavigationRetainedStateTest.kt
@@ -64,7 +64,7 @@ class GestureNavigationRetainedStateTest {
 
       setContent {
         CircuitCompositionLocals(circuit) {
-          val backStack = rememberSaveableBackStack { push(TestScreen.ScreenA) }
+          val backStack = rememberSaveableBackStack(TestScreen.ScreenA)
           val navigator =
             rememberCircuitNavigator(
               backStack = backStack,